example_auto_generate_ddl/requirements.txt
python-dotenv
faiss-cpu
psycopg
langchain
langchain-google-genai
langchain-community
langgraph
